+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Event admins cannot update surveys [#21](https://github.com/rokwire/surveys-building-block/issues/21)
+
 ## [1.2.0] - 2023-07-27
 ### Added
 - Associate surveys with Calendar BB events [#19](https://github.com/rokwire/surveys-building-block/issues/19)